Remove compatibility code for unsupported versions of Node (<16)
